+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/models/chores/gym/Exercise.kt
@@ -1,9 +1,8 @@
 package com.altmann.choresmanager.models.chores.gym
 
 class Exercise(
-    val exerciseId: Int,
+    val exerciseId: String,
     val name: String,
-    val sets : Int?,
-    val reps : Int?,
-    val restTimeSeconds : Int?
+    val sets: Int?,
+    val reps: Int?
 )