+++ composeApp/src/commonMain/kotlin/com/altmann/choresmanager/ui/screens/addchorepopup/AddChorePopup.kt
@@ -1,25 +1,10 @@
 package com.altmann.choresmanager.ui.screens.addchorepopup
 
 import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonColors
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
@@ -28,17 +13,21 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import com.altmann.choresmanager.models.chores.Priority
 import com.altmann.choresmanager.models.chores.Chore
+import com.altmann.choresmanager.models.chores.Priority
+import com.altmann.choresmanager.models.chores.gym.Exercise
+import com.altmann.choresmanager.ui.screens.addchorepopup.gymchore.GymChoreFields
 import com.altmann.choresmanager.ui.screens.components.fields.DateTextField
 import com.altmann.choresmanager.ui.screens.components.fields.TimeTextField
+import com.altmann.choresmanager.ui.screens.components.pickers.ChoreTypePicker
 import com.altmann.choresmanager.ui.screens.components.pickers.ColorPicker
 import com.altmann.choresmanager.ui.screens.components.pickers.DayOfWeekPicker
 import com.altmann.choresmanager.ui.screens.components.pickers.PriorityPicker
 import com.altmann.choresmanager.utils.DateTimeParser
-import kotlinx.datetime.DayOfWeek
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalTime
+import org.koin.compose.viewmodel.koinViewModel
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
 
@@ -47,20 +36,22 @@ fun AddChorePopup(
     onDismiss: () -> Unit,
     date: LocalDate,
     addChore: (chore: Chore) -> Unit,
-    visible: Boolean
+    visible: Boolean,
+    viewModel: ChorePopupViewModel
 ) {
-
+    val state by viewModel.state.collectAsState()
     DropdownMenu(
         expanded = visible,
         onDismissRequest = { onDismiss() },
         modifier = Modifier.width(300.dp).padding(horizontal = 8.dp),
         containerColor = MaterialTheme.colorScheme.background,
         border = BorderStroke(0.5.dp, MaterialTheme.colorScheme.primary),
-    )
-    {
+    ) {
         PopUpContent(
             date = date,
-            addChore = addChore,
+            state = state,
+            onEvent = viewModel::onEvent,
+            onAddChore = addChore
         )
     }
 }
@@ -69,97 +60,131 @@ fun AddChorePopup(
 @Composable
 fun PopUpContent(
     date: LocalDate,
-    addChore: (chore: Chore) -> Unit,
+    state: ChorePopupState,
+    onEvent: (ChorePopupEvent) -> Unit,
+    onAddChore: (chore: Chore) -> Unit,
 ) {
-    val title = remember { mutableStateOf("") }
-    val startDateTxt =
-        remember { mutableStateOf(TextFieldValue(DateTimeParser.parseDateToText(date))) }
-    val endDateTxt = remember { mutableStateOf(TextFieldValue("")) }
-    val startTimeTxt = remember { mutableStateOf(TextFieldValue("")) }
-    val endTimeTxt = remember { mutableStateOf(TextFieldValue("")) }
-    var selectedDays by remember { mutableStateOf<List<DayOfWeek>>(emptyList()) }
-    val selectedPriority = remember { mutableStateOf(Priority.NORMAL) }
-    val selectedColor = remember { mutableStateOf(Color.Blue) }
-
-    var openColorSelection by remember { mutableStateOf(false) }
-
     val startDate = remember { mutableStateOf(date) }
     val endDate = remember { mutableStateOf(LocalDate(2024, 6, 1)) }
     val startTime = remember { mutableStateOf(LocalTime(12, 0)) }
     val endTime = remember { mutableStateOf(LocalTime(14, 0)) }
 
+    val startDateTxt =
+        remember { mutableStateOf(TextFieldValue(DateTimeParser.parseDateToText(date))) }
+    val endDateTxt = remember { mutableStateOf(TextFieldValue("")) }
+    val startTimeTxt = remember { mutableStateOf(TextFieldValue("")) }
+    val endTimeTxt = remember { mutableStateOf(TextFieldValue("")) }
+
+
+
     Column(horizontalAlignment = Alignment.CenterHorizontally) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            Text(
+                "Add Chore",
+                style = MaterialTheme.typography.titleMedium,
+                modifier = Modifier.padding(8.dp)
+            )
+
+            Spacer(modifier = Modifier.weight(1f))
+
+            ChoreTypePicker(
+                selectedChore = state.selectedChoreType,
+                onChoreSelected = { onEvent(ChorePopupEvent.ChoreTypeChanged(it)) },
+                modifier = Modifier.padding(8.dp)
+            )
+        }
+
         TextField(
-            value = title.value,
-            onValueChange = { title.value = it },
+            value = state.title,
+            onValueChange = { onEvent(ChorePopupEvent.TitleChanged(it)) },
             maxLines = 1,
             label = { Text("Chore Title") },
         )
+
         FieldSpacer()
-        DateTextField(
-            dateTxt = startDateTxt,
-            date = startDate,
-            label = "Start date",
-            modifier = Modifier
-        )
+
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            DateTextField(
+                dateTxt = startDateTxt,
+                date = startDate,
+                label = "Start date",
+                modifier = Modifier.weight(1f)
+            )
+            Spacer(modifier = Modifier.width(4.dp))
+            DateTextField(
+                dateTxt = endDateTxt,
+                date = endDate,
+                label = "End date",
+                modifier = Modifier.weight(1f)
+            )
+        }
         FieldSpacer()
-        DateTextField(
-            dateTxt = endDateTxt,
-            date = endDate,
-            label = "End date",
-            modifier = Modifier
-        )
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            TimeTextField(
+                timeTxt = startTimeTxt,
+                time = startTime,
+                label = "Start Time",
+                modifier = Modifier.weight(1f)
+            )
+
+            Spacer(modifier = Modifier.width(4.dp))
+
+            TimeTextField(
+                timeTxt = endTimeTxt,
+                time = endTime,
+                label = "End Time",
+                modifier = Modifier.weight(1f)
+            )
+        }
+
         FieldSpacer()
-        TimeTextField(
-            timeTxt = startTimeTxt,
-            time = startTime,
-            label = "Start Time",
-            modifier = Modifier
-        )
-        FieldSpacer()
-        TimeTextField(
-            timeTxt = endTimeTxt,
-            time = endTime,
-            label = "End Time",
-            modifier = Modifier
-        )
-        FieldSpacer()
+
+        when (state.selectedChoreType) {
+            "Default" -> {}
+            "College" -> CollegeChoreFields(onEvent, state)
+            "Grocery" -> GroceryChoreFields(onEvent, state)
+            "Gym" -> GymChoreFields(onEvent, state)
+        }
+
         Text(
             "Days of Week",
             style = MaterialTheme.typography.bodySmall,
             modifier = Modifier.padding(start = 8.dp)
         )
+
         DayOfWeekPicker(
-            onSelect = { day ->
-                selectedDays = if (selectedDays.contains(day)) {
-                    selectedDays - day
-                } else {
-                    selectedDays + day
-                }
-            }, selectedDays = selectedDays
+            onSelect = { day -> onEvent(ChorePopupEvent.DayToggled(day)) },
+            selectedDays = state.selectedDays
         )
+
         Row(
             modifier = Modifier.fillMaxWidth(),
             verticalAlignment = Alignment.Bottom
         ) {
             PriorityPicker(
-                onPrioritySelected = { selectedPriority.value = it },
-                selectedPriority = selectedPriority.value
+                onPrioritySelected = { prio ->
+                    onEvent(ChorePopupEvent.PriorityChanged(prio.name))
+                },
+                selectedPriority = Priority.valueOf(state.selectedPriority.uppercase())
             )
+
             Spacer(modifier = Modifier.weight(1f))
+
             Column(horizontalAlignment = Alignment.CenterHorizontally) {
                 Text(
                     "Chore Color",
                     style = MaterialTheme.typography.bodySmall,
                 )
                 Button(
-                    onClick = {
-                        openColorSelection = !openColorSelection
-                    },
+                    onClick = { onEvent(ChorePopupEvent.ToggleColorPicker) },
                     shape = RoundedCornerShape(8.dp),
                     modifier = Modifier.width(100.dp),
-                    colors = ButtonColors(
-                        containerColor = selectedColor.value,
+                    colors = ButtonDefaults.buttonColors(
+                        containerColor = state.selectedColor,
                         contentColor = Color.White,
                         disabledContentColor = Color.Gray,
                         disabledContainerColor = Color.LightGray
@@ -168,44 +193,95 @@ fun PopUpContent(
                     Text("")
                 }
             }
+
             DropdownMenu(
-                expanded = openColorSelection,
-                onDismissRequest = { openColorSelection = false },
+                expanded = state.isColorPickerOpen,
+                onDismissRequest = { onEvent(ChorePopupEvent.ToggleColorPicker) },
                 modifier = Modifier.padding(horizontal = 8.dp),
                 offset = DpOffset(x = 100.dp, y = 0.dp),
                 containerColor = MaterialTheme.colorScheme.background,
             ) {
                 ColorPicker(
-                    onColorSelected = {
-                        selectedColor.value = it
-                    },
-                    selectedColor = selectedColor.value
+                    onColorSelected = { color -> onEvent(ChorePopupEvent.ColorChanged(color)) },
+                    selectedColor = state.selectedColor
                 )
             }
         }
+
         Button(
             onClick = {
-                addChore(
-                    Chore(
-                        choreId = Uuid.random().toString(),
-                        startTime = startTime.value,
-                        endTime = endTime.value,
-                        daysOfWeek = selectedDays,
-                        startDate = startDate.value,
-                        endDate = endDate.value,
-                        choreException = listOf(),
-                        title = title.value,
-                        description = "Treinão de perna",
-                        priority = selectedPriority.value,
-                        color = selectedColor.value
-                    )
+                val chore = Chore(
+                    choreId = Uuid.random().toString(),
+                    startTime = startTime.value,
+                    endTime = endTime.value,
+                    daysOfWeek = state.selectedDays,
+                    startDate = startDate.value,
+                    endDate = endDate.value,
+                    choreException = listOf(),
+                    title = state.title,
+                    description = state.description.ifBlank { "" },
+                    priority = Priority.valueOf(state.selectedPriority.uppercase()),
+                    color = state.selectedColor
                 )
-                print(date.toString() + "\n")
+                onEvent(ChorePopupEvent.AddChore(chore))
+                onAddChore(chore)
             },
             shape = RoundedCornerShape(8.dp),
             modifier = Modifier.clip(RoundedCornerShape(8.dp))
         ) {
             Text("Add Chore")
+        }
+    }
+}
+
+
+
+@Composable
+fun GroceryChoreFields(onEvent: (ChorePopupEvent) -> Unit, state: ChorePopupState) {
+    TODO("Not yet implemented")
+}
+
+@Composable
+fun CollegeChoreFields(onEvent: (ChorePopupEvent) -> Unit, state: ChorePopupState) {
+    Column {
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            TextField(
+                value = state.subject,
+                onValueChange = { subject -> onEvent(ChorePopupEvent.SubjectChanged(subject)) },
+                maxLines = 1,
+                modifier = Modifier.weight(1f),
+                label = { Text("Subject") },
+            )
+            Spacer(modifier = Modifier.width(4.dp))
+            TextField(
+                value = state.location,
+                onValueChange = { location -> onEvent(ChorePopupEvent.LocationChanged(location)) },
+                maxLines = 1,
+                modifier = Modifier.weight(1f),
+                label = { Text("Location") },
+            )
+        }
+        FieldSpacer()
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            TextField(
+                value = state.professor,
+                onValueChange = { subject -> onEvent(ChorePopupEvent.ProfessorChanged(subject)) },
+                maxLines = 1,
+                modifier = Modifier.weight(1f),
+                label = { Text("Professor") },
+            )
+            Spacer(modifier = Modifier.width(4.dp))
+            TextField(
+                value = state.totalHours,
+                onValueChange = { location -> onEvent(ChorePopupEvent.TotalHoursChanged(location)) },
+                maxLines = 1,
+                modifier = Modifier.weight(1f),
+                label = { Text("Total hours") },
+            )
         }
     }
 }